third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/dialog-open-expected.txt

Tests that dialog is visible after show() is called and not visible after close() is called. bug 90931

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle.getPropertyValue('display') is 'none'
PASS computedStyle.getPropertyValue('display') is 'block'
PASS computedStyle.getPropertyValue('display') is 'none'
PASS dialog.close(); threw exception InvalidStateError: Failed to execute 'close' on 'HTMLDialogElement': The element does not have an 'open' attribute, and therefore cannot be closed..
PASS successfullyParsed is true

TEST COMPLETE
